Addressed PR comments. APIs 'for later' have been removed.
