+++ ledger/ledger-utxo/src/main/kotlin/net/corda/v5/ledger/utxo/Command.kt
@@ -1,0 +1,9 @@
+package net.corda.v5.ledger.utxo
+
+import net.corda.v5.base.annotations.CordaSerializable
+
+/**
+ * Defines a marker interface which must be implemented by all [Contract] commands.
+ */
+@CordaSerializable
+interface Command
