--- ledger/ledger-utxo/src/test/java/net/corda/v5/ledger/utxo/BearableStateJavaApiTests.java
@@ -1,22 +1,0 @@
-package net.corda.v5.ledger.utxo;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-import java.security.PublicKey;
-import java.util.Set;
-
-public final class BearableStateJavaApiTests extends AbstractMockTestHarness {
-
-    @Test
-    public void getParticipantsShouldReturnTheExpectedValue() {
-        Set<PublicKey> value = bearableState.getParticipants();
-        Assertions.assertEquals(keys, value);
-    }
-
-    @Test
-    public void getIssuerShouldReturnTheExpectedValue() {
-        PublicKey value = bearableState.getBearer();
-        Assertions.assertEquals(bobKey, value);
-    }
-}
